Java Docs added to all Entity clases
Also changed access modifiers for User Login classes - And Changed datatype for ID from Integer to Long in ScanSummary Entity

+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/dsr/domain/file_integrity_scanner/entities/scan_summary/ScanSummary.java
@@ -13,56 +13,114 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
+/**
+ * This entity represents a summary of a scan performed on a file.
+ * It maps to a table in the database and establishes
+ * relationships
+ * with Scan, File, and Checksum entities through foreign keys.
+ */
 @Entity
 @Table(name = "scan_summary")
 public class ScanSummary extends PWSSbaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
+    /**
+     * Many-to-one relationship to the Scan entity.
+     * This field represents the scan that this summary belongs to.
+     */
     @ManyToOne(optional = false)
     @JoinColumn(name = "scan_id", nullable = false)
     private Scan scan;
 
+    /**
+     * Many-to-one relationship to the File entity.
+     * This field represents the file that was scanned.
+     */
     @ManyToOne(optional = false)
     @JoinColumn(name = "file_id", nullable = false)
     private File file;
 
+    /**
+     * Many-to-one relationship to the Checksum entity.
+     * This field represents the checksum of the scanned file.
+     */
     @ManyToOne(optional = false)
     @JoinColumn(name = "checksum_id", nullable = false)
     private Checksum checksum;
 
     // Getters and setters
 
-    public Integer getId() {
+    /**
+     * Gets the unique identifier of this scan summary entry.
+     *
+     * @return the id
+     */
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    /**
+     * Sets the unique identifier for this scan summary entry.
+     *
+     * @param id the id to set
+     */
+    public void setId(Long id) {
         this.id = id;
     }
 
+    /**
+     * Gets the Scan entity associated with this scan summary.
+     *
+     * @return the scan
+     */
     public Scan getScan() {
         return scan;
     }
 
+    /**
+     * Sets the Scan entity for this scan summary.
+     *
+     * @param scan the scan to set
+     */
     public void setScan(Scan scan) {
         this.scan = scan;
     }
 
+    /**
+     * Gets the File entity associated with this scan summary.
+     *
+     * @return the file
+     */
     public File getFile() {
         return file;
     }
 
+    /**
+     * Sets the File entity for this scan summary.
+     *
+     * @param file the file to set
+     */
     public void setFile(File file) {
         this.file = file;
     }
 
+    /**
+     * Gets the Checksum entity associated with this scan summary.
+     *
+     * @return the checksum
+     */
     public Checksum getChecksum() {
         return checksum;
     }
 
+    /**
+     * Sets the Checksum entity for this scan summary.
+     *
+     * @param checksum the checksum to set
+     */
     public void setChecksum(Checksum checksum) {
         this.checksum = checksum;
     }
